refactor: move embedding types to tensorzero-inference-types
Move EmbeddingProvider trait, Embedding, EmbeddingInput, EmbeddingRequest,
EmbeddingProviderResponse, EmbeddingProviderRequest, and
EmbeddingProviderRequestInfo from tensorzero-core to tensorzero-inference-types.

The RateLimitedRequest and RateLimitedResponse impls stay in core (orphan
rule allows this since the traits are core-local). This decouples 4
embedding providers (OpenAI, Azure, OpenRouter, Dummy) from the core
embeddings module.
